--- app/src/main/java/com/example/fanhy/myapplication/adapter/ImagePreviewVpAdapter.java
@@ -1,42 +1,0 @@
-package com.example.fanhy.myapplication.adapter;
-
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.view.ViewGroup;
-
-import java.util.List;
-
-/**
- * @anthor FanHY
- * @time 2018/8/29
- * @describe image预览total列表的adapter
- */
-public class ImagePreviewVpAdapter extends FragmentPagerAdapter {
-    private List<Fragment> mFragmentList;
-
-    public ImagePreviewVpAdapter(FragmentManager fm, List<Fragment> fragmentList) {
-        super(fm);
-        mFragmentList = fragmentList;
-    }
-
-    @Override
-    public Fragment getItem(int position) {
-        return mFragmentList.get(position);
-    }
-
-    @Override
-    public int getCount() {
-        return mFragmentList == null ? 0 : mFragmentList.size();
-    }
-
-    @Override
-    public Object instantiateItem(ViewGroup container, int position) {
-        return super.instantiateItem(container, position);
-    }
-
-    @Override
-    public void destroyItem(ViewGroup container, int position, Object object) {
-        super.destroyItem(container, position, object);
-    }
-}
