--- app/src/main/res/layout/activity_image_preview.xml
@@ -1,96 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activity.AlbumPreviewActivity">
-
-    <android.support.v4.view.ViewPager
-        android:id="@+id/vp_preview"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <include layout="@layout/head_album_activity" />
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_weight="1"
-            android:background="@color/transparent" />
-
-        <LinearLayout
-            android:id="@+id/ll_foot_preview"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:clickable="true"
-            android:orientation="vertical">
-
-            <android.support.v7.widget.RecyclerView
-                android:id="@+id/rv_preview"
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/album_preview_rv_height"
-                android:background="@color/transparentAA" />
-
-            <View
-                android:layout_width="match_parent"
-                android:layout_height="0.2dp"
-                android:background="@color/transparent70" />
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="@dimen/album_choose_head_foot_height"
-                android:layout_gravity="center_horizontal"
-                android:background="@color/transparentCC"
-                android:orientation="horizontal">
-
-                <View
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1" />
-
-                <CheckBox
-                    android:id="@+id/cbx_ori_picture"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:background="@null"
-                    android:button="@null"
-                    android:drawableLeft="@drawable/ori_pic_state"
-                    android:drawablePadding="20dp"
-                    android:gravity="center"
-                    android:paddingLeft="20dp"
-                    android:paddingRight="20dp"
-                    android:text="@string/ori_picture"
-                    android:textColor="@color/white" />
-
-                <View
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1" />
-
-                <CheckBox
-                    android:id="@+id/cbx_choose"
-                    android:layout_width="wrap_content"
-                    android:layout_height="match_parent"
-                    android:background="@null"
-                    android:button="@null"
-                    android:drawableLeft="@drawable/ori_pic_state"
-                    android:drawablePadding="20dp"
-                    android:gravity="center"
-                    android:paddingLeft="20dp"
-                    android:paddingRight="20dp"
-                    android:text="@string/choose"
-                    android:textColor="@color/white" />
-
-                <View
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="1" />
-            </LinearLayout>
-        </LinearLayout>
-    </LinearLayout>
-</FrameLayout>
