+++ picklocalpicture/src/main/res/drawable/image_preview_border.xml
