+++ picklocalpicture/src/main/res/drawable/ori_pic_state.xml
