+++ picklocalpicture/src/main/res/layout/activity_album_preview.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context="com.example.fanhy.picklocalpicture.activity.AlbumPreviewActivity">
+
+    <include layout="@layout/head_album_activity" />
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_weight="1">
+
+        <android.support.v7.widget.RecyclerView
+            android:id="@+id/rv_choose"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:background="@color/black" />
+
+        <ImageView
+            android:id="@+id/iv_mask"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:background="@color/transparent70"
+            android:visibility="gone" />
+
+        <android.support.v7.widget.RecyclerView
+            android:id="@+id/rv_switch_folder"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="bottom"
+            android:background="@color/white"
+            android:padding="10dp"
+            android:visibility="invisible" />
+    </FrameLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/album_choose_head_foot_height"
+        android:background="@color/transparentCC"
+        android:orientation="horizontal">
+
+        <Button
+            android:id="@+id/btn_switch_folder"
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="1"
+            android:background="@null"
+            android:gravity="center_vertical|left"
+            android:paddingLeft="20dp"
+            android:textColor="@color/white" />
+
+        <CheckBox
+            android:id="@+id/cbx_ori_picture"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:background="@null"
+            android:button="@null"
+            android:drawableLeft="@drawable/ori_pic_state"
+            android:drawablePadding="20dp"
+            android:gravity="center"
+            android:text="@string/ori_picture"
+            android:textColor="@color/white" />
+
+        <Button
+            android:id="@+id/btn_preview"
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="1"
+            android:background="@null"
+            android:gravity="center_vertical|right"
+            android:paddingRight="20dp"
+            android:text="@string/preview"
+            android:textColor="@color/white" />
+    </LinearLayout>
+
+</LinearLayout>
