+++ picklocalpicture/src/main/res/layout/head_album_activity.xml
