+++ picklocalpicture/src/main/res/layout/item_folder_switch.xml
