+++ picklocalpicture/src/main/res/layout/item_image_choose.xml
