+++ picklocalpicture/src/main/res/values/colors.xml
