+++ picklocalpicture/src/main/res/values/dimens.xml
