+++ picklocalpicture/src/main/res/values/styles.xml
